Split gd_color.c[ch] (sync with upstream)
