+++ ext/gd/libgd/gd_crop.c
@@ -24,9 +24,9 @@
 #include <math.h>
 
 #include "gd.h"
+#include "gd_color.h"
 
 static int gdGuessBackgroundColorFromCorners(gdImagePtr im, int *color);
-static int gdColorMatch(gdImagePtr im, int col1, int col2, float threshold);
 
 /**
  * Function: gdImageCrop
@@ -291,27 +291,3 @@ static int gdGuessBackgroundColorFromCorners(gdImagePtr im, int *color)
 		return 0;
 	}
 }
-
-static int gdColorMatch(gdImagePtr im, int col1, int col2, float threshold)
-{
-	const int dr = gdImageRed(im, col1) - gdImageRed(im, col2);
-	const int dg = gdImageGreen(im, col1) - gdImageGreen(im, col2);
-	const int db = gdImageBlue(im, col1) - gdImageBlue(im, col2);
-	const int da = gdImageAlpha(im, col1) - gdImageAlpha(im, col2);
-	const int dist = dr * dr + dg * dg + db * db + da * da;
-
-	return (100.0 * dist / 195075) < threshold;
-}
-
-/*
- * To be implemented when we have more image formats.
- * Buffer like gray8 gray16 or rgb8 will require some tweak
- * and can be done in this function (called from the autocrop
- * function. (Pierre)
- */
-#if 0
-static int colors_equal (const int col1, const in col2)
-{
-
-}
-#endif
